requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
scipy>=1.7.0
scikit-image>=0.19.0
opencv-python>=4.5.0
cryptography>=3.4.0  # Added for Fernet encryption/decryption
questionary==2.0.1
rich>=10.0.0
reedsolo>=1.5.0